python_mysql/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
mysql-connector==2.2.9
mysql-connector-python==8.0.29
protobuf==3.20.1
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
urllib3==1.26.9
zcrmsdk==2.0.12
